# [Feature][FlatJson] support type inference on flat json