… be/src/exprs/jsonpath.h
… be/src/storage/rowset/column_reader.cpp
… be/src/storage/rowset/json_column_iterator.h
… be/src/storage/rowset/json_column_writer.cpp
… be/src/util/CMakeLists.txt
… be/src/util/json_converter.h
… be/src/util/json_flater.cpp
… be/src/util/json_flater.h
… be/src/util/json_util.cpp
… be/src/util/json_util.h
… be/test/CMakeLists.txt
… be/test/exprs/flat_json_functions_test.cpp
… be/test/exprs/json_functions_test.cpp
… be/test/storage/rowset/flat_json_column_rw_test.cpp
… fe/fe-core/src/main/java/com/starrocks/catalog/ColumnAccessPath.java
… fe/fe-core/src/main/java/com/starrocks/catalog/FunctionSet.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rewrite/scalar/SimplifiedPredicateRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/tree/prunesubfield/PruneSubfieldRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/tree/prunesubfield/SubfieldAccessPathNormalizer.java
… fe/fe-core/src/main/java/com/starrocks/sql/plan/PlanFragmentBuilder.java
… fe/fe-core/src/test/java/com/starrocks/sql/plan/ExpressionTest.java
… fe/fe-core/src/test/java/com/starrocks/sql/plan/JsonTypeTest.java
… fe/fe-core/src/test/java/com/starrocks/sql/plan/LowCardinalityArrayTest.java
… fe/fe-core/src/test/java/com/starrocks/sql/plan/PruneComplexSubfieldTest.java
… gensrc/script/functions.py
… gensrc/thrift/PlanNodes.thrift
… test/sql/test_semi/R/test_flat_json_with_type
… test/sql/test_semi/T/test_flat_json_with_type
… test/sql/test_semi/T/test_invalid_flat_json