## Why I'm doing:
Support type inference on flat json

## What I'm doing:
### write process
In write process, do type inference when json flating
Support vpack json type level inference:
None -> SmallInt -> Int -> String -> Json

the mapping type in sr:
bool ->  SmallInt -> BigInt -> String -> Json

will auto cast to compatible type in write process.

(Don't support json-bool, because sr-bool show 1 or 0, but json-bool will show as True or False, it's will take some function result different)

![image](https://github.com/StarRocks/starrocks/assets/5609319/71c724a2-726f-4196-a1b5-ee9269d17bda)


### read process
In read process, FE do type inference by SQL(json function), and auto cast the flat type by required
the process like:

![image](https://github.com/StarRocks/starrocks/assets/5609319/41649175-77f6-42c2-80ba-768024a881a0)

other update:
1. move JsonFlater class from JsonUtil.h to JsonFlater.h
2. refactor FlatJsonWrite logical, extract type inference to JsonFlater
3. json function update:
    a. support `get_json_bool`: to read json bool
    b. update `get_json_int`: return int to return bigint, vpack json int is Bigint at actual, not int
    c. update `get_json_xxx(varchar, varchar)` logical: to keep same as `get_json_xxx(json, varchar)`
    d. update `get_json_key`: to support flat json version 
4. optimize json function prepare performance
    a. add `NativeJsonState` to handle path inference and cast expr only once
5. update FE `pruneSubfieldRule`:
    a. to inerence json read compatible type
    b. to optimize `cast(json to xxx)`  to `get_json_xxx`, to use flat json
6. FE fold `->` operator and `json_query` function
7. `ColumnAccessPath`: add type_desc to descibe the leaf node type 
8. update cast_from_json:
    a. follow Mysql, support json-boolean cast to number (cast to 1/0)
    b. follow Mysql, support json-number cast to boolean (check != 0)
9. update json cast/get_json_xx/get_native_json_xx logical, keep same as `get_json_xxx`

### preformance test
same as #37577, 100G ssb

#### write
Load Json with Json Flatten Time cost(this PR): 1303.611s
Load Json with Json Flatten Time cost(before): 1640.965s
Load Json without Json Flatten Time cost: 790.653s
Load struct Time cost: 276.334s
Load origin data Time costs: 247.103s

#### read
| SSB    | origin | Json     | FlatJson(Before) | FlatJson(This PR) | Struct(Int) |
| ----   | ----   | ----     | ----             | ----           | -----       |
|  Q01   |  84    | 18895    |  5589            |  1533          | 374         |
|  Q03   |  64    | 9553     |  4465            |  888           |  180        |
|  Q04   |  933   | 319234   |  109882          |  18006         |  3959       |
|  Q05   |  668   | 319833   |  55168           |  17941         |  3956       |
|  Q06   |  503   | 319020   |  44419           |  13369         |   2822      |
|  Q07   |  790   | 319020   |  129003          |  18075         |  3862       |
              
#### data usage
```
-- lineorder is origin data
-- lineorder_flat_js is flat json(This)
-- lineorder_flat_js1 is flat json(Before)
-- lineorder_js is json
+-----------------------------+----------------+---------------------+
| TableName                   | Size           | ReplicaCount        |
+-----------------------------+----------------+---------------------+
| lineorder_flat              | 58.827 GB      | 336                 | 
| lineorder_flat_js(This)     | 196.918 GB     | 336                 | 
| lineorder_flat_js1(Before)  | 292.569 GB     | 336                 |
| lineorder_js                | 141.805 GB     | 336                 |
+-----------------------------+----------------+---------------------+
```
### Follow #37577

### PR split
this PR is too large, split to different PR to review and merge:
- [x] refactor json cast logical: https://github.com/StarRocks/starrocks/pull/43503
- [x] fold json_query rule: https://github.com/StarRocks/starrocks/pull/43504
- [ ] FE support inerence json type: https://github.com/StarRocks/starrocks/pull/43526
- [ ] BE JsonFlater core: https://github.com/StarRocks/starrocks/pull/43528
- [x] BE add new get_json_xx function: https://github.com/StarRocks/starrocks/pull/43547

## What type of PR is this:

- [ ] BugFix
- [x] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5